@@ -7,10 +7,8 @@ namespace DoTheSpriteThing.FileProcessors
     {
         public void CreateSprite(MagickImageCollection spriteImages, string spriteFilename)
         {
-            using (MagickImage result = spriteImages.AppendVertically())
-            {
-                result.Write(spriteFilename);
-            }
+            using var result = spriteImages.AppendVertically();
+            result.Write(spriteFilename);
         }
     }
 }
